Version 1.4.2
### Features
- ITestLocal - Run all projects with `*.Tests` only Local Build
- ITestServer - Run all projects with `*.Tests` only Server Build
### Updated
- Set `Nuke.Common` Version = `6.2.1`
### Added
- Add `IHazTest` with base tests methods.
